+++ dev-packages/application-manager/src/generator/frontend-generator.ts
@@ -111,7 +111,7 @@ const electron = require('electron');
 const { join, resolve } = require('path');
 const { isMaster } = require('cluster');
 const { fork } = require('child_process');
-const { app, BrowserWindow, ipcMain, Menu } = electron;
+const { app, shell, BrowserWindow, ipcMain, Menu } = electron;
 
 const applicationName = \`${this.pck.props.frontend.config.applicationName}\`;
 
@@ -124,9 +124,6 @@ if (isMaster) {
             role: 'help', submenu: [{ role: 'toggledevtools'}]
         }]));
 
-        // Window list tracker.
-        const windows = [];
-
         function createNewWindow(theUrl) {
 
             // We must center by hand because \`browserWindow.center()\` fails on multi-screen setups
@@ -138,40 +135,25 @@ if (isMaster) {
             const y = Math.floor(bounds.y + (bounds.height - height) / 2);
             const x = Math.floor(bounds.x + (bounds.width - width) / 2);
 
-            const newWindow = new BrowserWindow({ width, height, x, y, show: !!theUrl, title: applicationName });
+            // Always hide the window, we will show the window when it is ready to be shown in any case.
+            const newWindow = new BrowserWindow({ width, height, x, y, show: false, title: applicationName });
+            newWindow.on('ready-to-show', () => newWindow.show());
 
-            if (windows.length === 0) {
-                newWindow.webContents.on('new-window', (event, url, frameName, disposition, options) => {
-                    // If the first electron window isn't visible, then all other new windows will remain invisible.
-                    // https://github.com/electron/electron/issues/3751
-                    options.show = true;
-                    options.width = width;
-                    options.height = height;
-                    options.title = applicationName;
-                });
-            }
-            windows.push(newWindow);
+            // Prevent calls to "window.open" from opening an ElectronBrowser window,
+            // and rather open in the OS default web browser.
+            newWindow.webContents.on('new-window', (event, url) => {
+                event.preventDefault();
+                shell.openExternal(url);
+            });
+
             if (!!theUrl) {
                 newWindow.loadURL(theUrl);
-            } else {
-                newWindow.on('ready-to-show', () => newWindow.show());
             }
-            newWindow.on('closed', () => {
-                const index = windows.indexOf(newWindow);
-                if (index !== -1) {
-                    windows.splice(index, 1);
-                }
-                if (windows.length === 0) {
-                    app.exit(0);
-                }
-            });
             return newWindow;
         }
 
         app.on('window-all-closed', () => {
-            if (process.platform !== 'darwin') {
-                app.quit();
-            }
+            app.quit();
         });
         ipcMain.on('create-new-window', (event, url) => {
             createNewWindow(url);
@@ -182,7 +164,9 @@ if (isMaster) {
         const devMode = process.defaultApp || /node_modules[\/]electron[\/]/.test(process.execPath);
         const mainWindow = createNewWindow();
         const loadMainWindow = (port) => {
-            mainWindow.loadURL('file://' + join(__dirname, '../../lib/index.html') + '?port=' + port);
+            if (!mainWindow.isDestroyed()) {
+                mainWindow.loadURL('file://' + join(__dirname, '../../lib/index.html') + '?port=' + port);
+            }
         };
 
         // We cannot use the \`process.cwd()\` as the application project path (the location of the \`package.json\` in other words)